requirements.txt
aws-cdk.aws-s3==1.30.0
aws-cdk.aws-s3-deployment==1.30.0
boto3